healthcheck: remove HealthCheck interface, rename HealthCheckImpl -> HealthCheck, move checkConn and finalizeConn to tabletHealthCheck, rename deleteConn -> deleteTablet
