+++ go/vt/vtgate/api.go
@@ -88,7 +88,7 @@ func getItemPath(url string) string {
 	return parts[1]
 }
 
-func initAPI(ctx context.Context, hc discovery.HealthCheck) {
+func initAPI(ctx context.Context, hc *discovery.HealthCheck) {
 	// Healthcheck real time status per (cell, keyspace, tablet type, metric).
 	handleCollection("health-check", func(r *http.Request) (interface{}, error) {
 		cacheStatus := hc.CacheStatus()
