+++ go/vt/vtgate/discoverygateway.go
@@ -411,6 +411,6 @@ func NewShardError(in error, target *querypb.Target, tablet *topodatapb.Tablet) 
 
 // HealthCheck should never be called on a DiscoveryGateway
 // This exists only to satisfy the interface
-func (dg *DiscoveryGateway) HealthCheck() discovery.HealthCheck {
+func (dg *DiscoveryGateway) HealthCheck() *discovery.HealthCheck {
 	return nil
 }
